// models/Room.ts
import {
    BelongsToManyAddAssociationMixin,
    BelongsToManyGetAssociationsMixin,
    CreationOptional,
    Model,
    InferAttributes,
    InferCreationAttributes,
    NonAttribute
} from 'sequelize';
import { sequelize, DataTypes } from '../utils/database';
import { UserModel } from './User';
import { MessageModel } from './Message';

export interface RoomModel extends Model<InferAttributes<RoomModel>, InferCreationAttributes<RoomModel>> {
    id: CreationOptional<string>,
    isGroup: boolean,
    roomName: string,
    createdAt: CreationOptional<Date>,
    updatedAt: CreationOptional<Date>,
    // mixin methods (these are auto-generated by sequelize we only define signatures for TS)
    addUsers: BelongsToManyAddAssociationMixin<UserModel, string[]>,
    getUsers: BelongsToManyGetAssociationsMixin<UserModel>,
    Users?: UserModel[],

    Messages?: MessageModel[]
}

const Room = sequelize.define<RoomModel>('Room', {
    id: {
        primaryKey: true,
        type: DataTypes.UUID,
        defaultValue: DataTypes.UUIDV4,
        allowNull: false
    },
    roomName: {
        type: DataTypes.STRING,
        allowNull: false
    },
    isGroup: {
        type: DataTypes.BOOLEAN,
        allowNull: false
    },
    createdAt: {
        type: DataTypes.DATE
    },
    updatedAt: {
        type: DataTypes.DATE
    },
}, {
    tableName: "rooms"
});

export default Room;